requirements.txt
Flask
Flask-CORS
torch
torchvision
numpy
Pillow